desktop: refactor collision and item placement
