+++ spooky-core/src/resources.rs
@@ -12,4 +12,5 @@ pub struct MazeResource {
 pub struct PlayerPosition {
     pub x: f32,
     pub y: f32,
+    pub z: f32,
 }
