+++ spooky-core/src/systems/mod.rs
@@ -1,3 +1,3 @@
 pub mod setup;
-pub mod player_input;
+pub mod process_player_input;
 pub mod game_logic;
